requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
django-cors-headers==3.11.0
django-filter==22.1